backend/requirements.txt
fastapi
uvicorn
python-socketio
pydantic
python-dotenv
firebase-admin
PyJWT
passlib
Pillow
google-cloud-firestore

twilio
